Question 6
I updated the OpportunityTrigger to look for before delete context.
I then added a method to the OpportunityTriggerHandler called beforeDelete.
In this method I check to see if the Opportunity is Closed Won and if the related Account Industry is 'Banking'.
If both are true then I throw an error to the User and stop them from deleting the Opportunity.